refactor(tangle-dapp): Simplify, fix errors
