+++ libs/tangle-shared-ui/src/context/RestakeContext/index.ts
@@ -1,4 +1,3 @@
 export { default as RestakeContext } from './RestakeContext';
 export { default as RestakeContextProvider } from './RestakeContextProvider';
-export * from './types';
 export { default as useRestakeContext } from './useRestakeContext';
