--- libs/tangle-shared-ui/src/data/rewards/useVaultPotAccount.ts
@@ -1,17 +1,0 @@
-import { useCallback } from 'react';
-import useApiRx from '../../hooks/useApiRx';
-import rewardVaultsPotAccountsRxQuery from '../../queries/restake/rewardVaultsPotAccounts';
-
-const useVaultPotAccount = () => {
-  return useApiRx(
-    useCallback((apiRx) => {
-      if (apiRx.query.rewards?.rewardVaultsPotAccount === undefined) {
-        return null;
-      }
-
-      return rewardVaultsPotAccountsRxQuery(apiRx);
-    }, []),
-  );
-};
-
-export default useVaultPotAccount;
